+++ iModelBridgeCore/Pointools/PointoolsVortexAPI/src/ptcloud2/scanimage.cpp
@@ -161,7 +161,7 @@ static float pix_overlap(double x, double y, double xd, double yd)
 							(x - 0.5) - (xd + 0.5);
 	double yc = (y < yd) ? (y + 0.5) - (yd - 0.5):
 							(y - 0.5) - (yd + 0.5);
-	return (xc * yc) > 0 ? (xc * yc) : 0;
+	return (xc * yc) > 0 ? static_cast<float>(xc * yc) : 0.0f;
 }
 //
 static int indices [] = {	-1,1,	0,1,	1,1,
